[Parley] refactor: Extract node handlers to MainWindow.NodeHandlers.cs (#1222)
Moved 8 node CRUD methods (~179 lines) from MainWindow.axaml.cs to new partial file.
Build succeeds, all 519 tests pass.
